rec_svc/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
torch>=2.1.0
transformers>=4.44.0
Pillow>=10.3.0
faiss-cpu>=1.8.0
numpy>=1.26.0
requests>=2.32.0
pydantic>=2.7.0
